Add files via upload
Here we present a basic and well-known example for the application of neural networks: recognising handwritten digits. We compare two standard methods known in machine learning, i.e. decision trees and neural networks, and we compare their performance. For neural networks, we use two approaches. The first is a simple neural network with no hidden layers. Then we add hidden layers to the network to observe the increase in the method's effectiveness.